Use regexes for jsonschema pre-validation
This commit adds the ability to dynamically pick the right
schema to validate the YAML data against, based on the apiVersion
included in the YAML metadata.

This commit also corrects the pre-validation schema by changing
"choices" to "pattern", allowing Deckhand to simply check for a
regex rather than having to enumerate all possible types of data.