Change .to_oslo_conf to .to_ini
Currently, openstackhelm-toolkit provides two utility methods:
.to_oslo_conf and .to_ini whose only difference at this time is
the ability to handle multistring in .to_oslo_conf as it is an
openstack concept that is not found in .ini files.

As logging.conf and paste.ini do not support multistring, they do
not need to be render by .to_oslo_conf, and while the code is not
incorrect, using .to_ini would be more appropriate.

Trivial fix

Change-Id: I20ae77842a3406ea64bf40d951adef2e519cbfcb