Add expected length validation to gabbi functional tests
This PS adds expected length validations to gabbi functional tests
in order for there to be more robust response body validations
for theses tests.

This is done via

  $.`len`: 1

under the attribute "response_multidoc_jsonpaths", which validates
that the returned list of items (we coerce everything into a list
for these tests, such that a {} is coerced into [{}]) is of expected
length 1.

Also break up static query parameters into dynamically generated ones
[0] to make the code more readable.

[0] Search for "query_parameters" in http://gabbi.readthedocs.io/en/latest/example.html

Change-Id: I8b1efc9a3a0276d7b383d0960f45b485bf89a234