[docs] Document schemas used for document validation
This PS adds documentation for all the document schemas that are
used for ``deckhand-schema-validation`` validations. This will
help users to know what their schemas need to look like for
Deckhand to accept them.

The documentation like most of it is generated automically via sphinx.
A docstring was provided under each module schema variable and
included inside __all__ because according to docs [0]:
"For modules, __all__ will be respected when looking for members;
the order of the members will also be the order in __all__."

[0] http://www.sphinx-doc.org/en/stable/ext/autodoc.html#directive-automodule

Change-Id: I224f3980dcf877bab400aa505728a332e67709cc