Optimize runtime for excluding deleted documents
Optimize runtime for excluding deleted documents in Deckhand's
database layer. Currently the runtime is O(N^2): For each
document, check to see if it is deleted, then iterate over
every other document and delete it if its creation date
is earlier than the current document's deletion date
and their schema and metadata.name are the same (in other
words delete every document from an earlier revision that
was deleted in a more current one if it's the same document).

The runtime was changed to O(NlogN).

Change-Id: I4aa4e1429014731751288861735c705e6b6c6ed4