Update Deckhand README
Part of the README is out of date and needs to be updated. The
Getting Started section was also updated to contain information
about how to run Deckhand via Docker as well as how to do a
complete manual install of Deckhand.

The testing documentation was also updated because some tox
jobs have been removed, so the commands in the README should too.

Change-Id: I4438d3b3462e06923969831242cb377237c03480