[TrivialFix] Log only if document parentSelector set
This is intended to lessen the number of LOG statements being
dumped out. The log statement regarding missing document
parent will only be issued if the document has a parentSelector
and yet no parent for it could be found. Also, the log level
is changed to DEBUG so that it's only issued if CONF.debug is
True.

Change-Id: I32d3a43604b5fce564f6af5579a0b173625533a1