Fix abstract parent documents substitutions not propagating
This is to fix abstract parent documents that have registered
substitutions but are not propagating those substitutions
down to concrete children that rely on that substituted data
(whereas it is irrelevant to the end user whether the abstract
parent receives those substitutions or not).

After this change abstract documents will always undergo
substitution such that their concrete children documents should
inherit those changes.

A unit test has been added for regression.

Change-Id: I73ed8d4caa6923492cc5ff042ecc57fbfeeb7c1c