Skip layering for control documents
This is to skip layering for control documents (those whose
metadata.schema starts with "deckhand/Control") as these documents
consist of ValidationPolicy or LayeringPolicy documents -- and
it would be both nonsensical and scary to try to layer
these types of documents.

Documentation for this will be updated during a larger overhaul
effort to improve Deckhand's documentation.

Change-Id: Ia785e54c4e26a4158b6bdc89da8b96b4455f7b39