[fix] Add uwsgi entrypoint options
Changes the entrypoint.sh options for uwsgi to include:
 -b 32768 : for larger header/url handling
 --die-on-term : for more 'normal'  handling of SIGTERM
 --lazy-apps : to delay init of python until after forking workers
 --master : to provide a master process for handling request dispatch

The purpose of these changes is intended to avoid some crash behavior
that is occuring when the process being forked has an open db connection.
The --lazy-apps option should delay initialization. The other options are
recommended by uwsgi documentation, specicially the --master option.
The larger buffer size is not strictly recommended, but matters when large
headers are included.

The die-on-term option should provide better behavior in the container
environment.

Related-Change: I60adeffff5461fdda957124232bc5a606baae413
Change-Id: I70510246576a8fb6aa216e7c9c7e97c1c9ab791c