Add integration tests job to .zuul.yaml
This patchset adds integration tests job to .zuul.yaml so that
they can be run against Zuul per each commit.

The number of jobs that are executed in the gate pipeline
have been pruned because only *-docker jobs are of critical
importance. We can safely skip over *-uwsgi jobs in order
to save on Zuul resources.

Change-Id: I98ce49163b2db64cbdf18fe00b82c99fecd4139d