Add missing Keystone options to registration of config
This patchset adds missing Keystone options to registration
of config, including timeout.

This fix is similar to a recent Shipyard patchset [0] which
aimed to solve the same issue but expands on it by following
Nova patterns here: [1].

This also updates the sample config file which now includes
the timeout option.

[0] https://review.openstack.org/#/c/572563/
[1] http://git.openstack.org/cgit/openstack/nova/tree/nova/conf/cinder.py

Change-Id: I4e52170e48dba7064c69f42f4cbbb7e54eae11a9