Add functional test for validating single source multi dest substitution
This patchset fixes a schema bug to allow for single source multi
dest substitution to work (it is currently rendering successfully
but failing on validation as the base Deckhand validation schema
was not updated).

This adds a functional test to avoid regression.

Also fixes failed_when condition for functional-tests
and integration-tests roles to check if 'commands failed'
appears in stdout which only happens when pytest runner
fails.

Change-Id: I9bb0c51b3374bbaf9fd34d5f93f1a1777bb3b1e8