layering: Support layering for primitive types
This patchset adds support for layering between primitive types
of data. It is not possible to merge together a string with
an integer or a boolean with a dictionary -- however, merge
actions can still be resolved in such cases in order to get layering
to work:

merge => child replaces parent data
replace => child replaces parent data
delete => empty dictionary (in effect)

The only path that is possible in such cases is "." as
a primitive doesn't have a JSON path. Else missing document key
exceptions are raised.

Change-Id: I9da557bf9444c37b3e59672f3a93f49fdf1e4a02