Remove deprecated substitution_sources kwarg
This patch set removes the deprecated `substitution_sources`
kwarg from DocumentLayering class. The necessary changes to
Pegleg and Promenade have been made here:

https://review.openstack.org/#/q/topic:substitution_sources+(status:open+OR+status:merged)

So this can be safely dropped.

Also note that this corrects a previously incorrect docstring
in secrets_manager: the substitution_sources kwarg there was
tagged as DEPRECATED but this is only true for the layering
module which currently serves as the front door into the
Deckhand engine.

Change-Id: Ia4872f237a70c9b0c322710bb1ac7db1079357bf