Invalidate rendered documents cache when deleting all revisions
When DELETE /revisions is called all revisions are deleted from
the DB, which means that all cached documents for each revision
should be cleared from memory. This patch set addresses this
issue. This is part of a set of patch sets aimed at fixing edge
cases like this so rendered documents are always returned correctly.

In the same vein, invalidation is performed for Barbican secrets,
for the same endpoint above.

Change-Id: I0334b638d6ce7569e8771e960dd283b1a85a5a6b