Support rolling back to revision 0
This patch set allows revision rollback functionality to work
with revision 0 in order to allow Deckhand to create a blank
revision (without any documents), without having to delete every
revision in the database (which is currently done here: [1] for
example). This is less than ideal as revision history is lost.
Instead, Deckhand should just support rolling back to revision 0
which simulates a blank slate (though in reality the previous
revision history is retained).

[1] https://github.com/openstack/airship-shipyard/blob/18ed6674d27ad80c87156effbd0f27d756c0bd8b/src/bin/shipyard_airflow/shipyard_airflow/control/helpers/configdocs_helper.py#L159

Change-Id: Ic94b25ca15f88ba492e08c907d4a330aeecca810