trivial: Fix error message for non-matching policy checks
This patch set ensures that the sets of the expected and actual
policies are dumped out for any failed policy check that happens
automatically within control tests. This is to make policy
mismatch errors a little easier to debug: For example, if expected
is ['foo', 'bar'] and actual is ['foo', 'foo', 'bar', 'baz'] it
means that the expected should be changed to ['foo', bar', 'baz']
but this is not immediately apparent without taking the set of
each list.

Change-Id: I3178e323ac780b2c3c81f7d9ba72aa6b171338f1