fix: Correct .data path layering edge case
This patch set corrects logic for an edge case in layering where
the action `path` is set to `.data`. In this case this means
that the root of the data section should be used, i.e. '.'
or '$.'. The previous adjustment was incorrect: .data was being
changed to empty string ''. This fixes that logic to change to
'.'.

Change-Id: Id6cf0d4d65020220c540eb162a33055035336cde