Make Deckhand integration tests non-voting
Deckhand integration tests are currently broken, and this makes them
non-voting.  The test harness is superflous now that the (also
non-voting) airskiff job is in place, and it should be considered as
follow-on work whether we should remove the integration tests in favor
of the airskiff integration job, or fix/maintain the current integration
test harness and/or tests themselves.

Change-Id: Ib3ce88693d80de6c9c79c278eed2e312fe3a0b6f